requirements.txt
streamlit
tensorflow==2.20.0
numpy
opencv-python-headless
Pillow